v5

+++ c/complex.c
@@ -35,10 +35,10 @@ int complex_s (float *Y, const float *X1, const float *X2, const size_t R1, cons
     }
     else if (iscolmajor)
     {
-        const int r1i = (int)(R1>1u), r2i = (int)(R2>1u);
-        const int c1i = (int)R1*((int)(C1>1u)-(int)(R1>1u)), c2i = (int)R2*((int)(C2>1u)-(int)(R2>1u));
-        const int s1i = (int)(R1*C1)*((int)(S1>1u)-(int)(C1>1u)), s2i = (int)(R2*C2)*((int)(S2>1u)-(int)(C2>1u));
-        const int h1i = (int)(R1*C1*S1)*((int)(H1>1u)-(int)(S1>1u)), h2i = (int)(R2*C2*S2)*((int)(H2>1u)-(int)(S2>1u));
+        const int r1i = (R1>1u), r2i = (R2>1u);
+        const int c1i = (int)R1*((C1>1u)-(R1>1u)), c2i = (int)R2*((C2>1u)-(R2>1u));
+        const int s1i = (int)(R1*C1)*((S1>1u)-(C1>1u)), s2i = (int)(R2*C2)*((S2>1u)-(C2>1u));
+        const int h1i = (int)(R1*C1*S1)*((H1>1u)-(S1>1u)), h2i = (int)(R2*C2*S2)*((H2>1u)-(S2>1u));
         for (size_t h=H; h>0u; --h, X1+=h1i, X2+=h2i)
         {
             for (size_t s=S; s>0u; --s, X1+=s1i, X2+=s2i)
@@ -55,10 +55,10 @@ int complex_s (float *Y, const float *X1, const float *X2, const size_t R1, cons
     }
     else
     {
-        const int h1i = (int)(H1>1u), h2i = (int)(H2>1u);
-        const int s1i = (int)H1*((int)(S1>1u)-(int)(H1>1u)), s2i = (int)H2*((int)(S2>1u)-(int)(H2>1u));
-        const int c1i = (int)(H1*S1)*((int)(C1>1u)-(int)(S1>1u)), c2i = (int)(H2*S2)*((int)(C2>1u)-(int)(S2>1u));
-        const int r1i = (int)(H1*S1*C1)*((int)(R1>1u)-(int)(C1>1u)), r2i = (int)(H2*S2*C2)*((int)(R2>1u)-(int)(C2>1u));
+        const int h1i = (H1>1u), h2i = (H2>1u);
+        const int s1i = (int)H1*((S1>1u)-(H1>1u)), s2i = (int)H2*((S2>1u)-(H2>1u));
+        const int c1i = (int)(H1*S1)*((C1>1u)-(S1>1u)), c2i = (int)(H2*S2)*((C2>1u)-(S2>1u));
+        const int r1i = (int)(H1*S1*C1)*((R1>1u)-(C1>1u)), r2i = (int)(H2*S2*C2)*((R2>1u)-(C2>1u));
         for (size_t r=R; r>0u; --r, X1+=r1i, X2+=r2i)
         {
             for (size_t c=C; c>0u; --c, X1+=c1i, X2+=c2i)
@@ -100,10 +100,10 @@ int complex_d (double *Y, const double *X1, const double *X2, const size_t R1, c
     }
     else if (iscolmajor)
     {
-        const int r1i = (int)(R1>1u), r2i = (int)(R2>1u);
-        const int c1i = (int)R1*((int)(C1>1u)-(int)(R1>1u)), c2i = (int)R2*((int)(C2>1u)-(int)(R2>1u));
-        const int s1i = (int)(R1*C1)*((int)(S1>1u)-(int)(C1>1u)), s2i = (int)(R2*C2)*((int)(S2>1u)-(int)(C2>1u));
-        const int h1i = (int)(R1*C1*S1)*((int)(H1>1u)-(int)(S1>1u)), h2i = (int)(R2*C2*S2)*((int)(H2>1u)-(int)(S2>1u));
+        const int r1i = (R1>1u), r2i = (R2>1u);
+        const int c1i = (int)R1*((C1>1u)-(R1>1u)), c2i = (int)R2*((C2>1u)-(R2>1u));
+        const int s1i = (int)(R1*C1)*((S1>1u)-(C1>1u)), s2i = (int)(R2*C2)*((S2>1u)-(C2>1u));
+        const int h1i = (int)(R1*C1*S1)*((H1>1u)-(S1>1u)), h2i = (int)(R2*C2*S2)*((H2>1u)-(S2>1u));
         for (size_t h=H; h>0u; --h, X1+=h1i, X2+=h2i)
         {
             for (size_t s=S; s>0u; --s, X1+=s1i, X2+=s2i)
@@ -120,10 +120,10 @@ int complex_d (double *Y, const double *X1, const double *X2, const size_t R1, c
     }
     else
     {
-        const int h1i = (int)(H1>1u), h2i = (int)(H2>1u);
-        const int s1i = (int)H1*((int)(S1>1u)-(int)(H1>1u)), s2i = (int)H2*((int)(S2>1u)-(int)(H2>1u));
-        const int c1i = (int)(H1*S1)*((int)(C1>1u)-(int)(S1>1u)), c2i = (int)(H2*S2)*((int)(C2>1u)-(int)(S2>1u));
-        const int r1i = (int)(H1*S1*C1)*((int)(R1>1u)-(int)(C1>1u)), r2i = (int)(H2*S2*C2)*((int)(R2>1u)-(int)(C2>1u));
+        const int h1i = (H1>1u), h2i = (H2>1u);
+        const int s1i = (int)H1*((S1>1u)-(H1>1u)), s2i = (int)H2*((S2>1u)-(H2>1u));
+        const int c1i = (int)(H1*S1)*((C1>1u)-(S1>1u)), c2i = (int)(H2*S2)*((C2>1u)-(S2>1u));
+        const int r1i = (int)(H1*S1*C1)*((R1>1u)-(C1>1u)), r2i = (int)(H2*S2*C2)*((R2>1u)-(C2>1u));
         for (size_t r=R; r>0u; --r, X1+=r1i, X2+=r2i)
         {
             for (size_t c=C; c>0u; --c, X1+=c1i, X2+=c2i)
